+++ CHANGELOG.md
@@ -1,6 +1,29 @@
 # Release Notes
 
-# 1.10.0 (2019-11-20)
+## 2.0.0 (2020-09-10)
+
+**Requires PlatformIO Core 5.0 or above**
+**Requires VSCode 1.44 or above**
+
+* New PlatformIO Task Explorer
+  - Instant access to the Project Tasks within the VSCode Explorer
+  - Grouped tasks: Generic, Advanced, PIO Remote, Platform, Custom, etc.
+  - Support for PlatformIO dev-platform tasks (Program FPGA bitstream, Burn bootloader, Upload to FS, OTA Update, etc). The list of tasks depends on a particular dev-platform
+  - Access to [User Custom Targets](https://docs.platformio.org/en/latest/projectconf/advanced_scripting.html#custom-targets)
+* New Project Environment Switcher
+  - Switch between project environments declared in [platformio.ini](https://docs.platformio.org/en/latest/projectconf/index.html) project configuration file (issue [#544](https://github.com/platformio/platformio-vscode-ide/issues/544))
+  - Activate IntelliSense service based on the current environment
+  - Automatically generate a debugging configuration for the active environment
+* New PlatformIO IDE Installer
+  - Added progress information
+  - Switched to the cross-platform and portable [get-platformio.py](https://github.com/platformio/platformio-core-installer) installer script
+  - Use built-in portable Python 3 on Windows and macOS (it can be disabled with ``platformio-ide.useBuiltinPython`` setting)
+* Added support for ``extends`` option in ``platformio.ini`` project configuration file (issue [#1371](https://github.com/platformio/platformio-vscode-ide/issues/1371))
+* Contribute PlatformIO Core CLI into VSCode's default Terminal
+* Reduced startup time (PlatformIO Core verification process)
+* Fixed a bug when hotkeys in PlatformIO Home did not work on macOS (issue [#606](https://github.com/platformio/platformio-vscode-ide/issues/606))
+
+## 1.10.0 (2019-11-20)
 
 - Minimal requirements for PlatformIO Core is >=4.1.0
 - Added PIO Check project task
@@ -12,16 +35,16 @@
     * Check Python's "urllib" SSL module
     * Skip Python from msys, mingw, emacs installations (issue [#1353](https://github.com/platformio/platformio-vscode-ide/issues/1353))
 
-# 1.9.3 (2019-10-29)
+## 1.9.3 (2019-10-29)
 
 - Handle "openTextDocument" in VSCode from upcoming PIO Home 3.0 and Project Inspect
 - Use single PIO Home Server instance per multiple windows/sessions
 
-# 1.9.2 (2019-10-20)
+## 1.9.2 (2019-10-20)
 
 - Fixed an issue with broken PlatformIO Core installation
 
-# 1.9.1 (2019-10-20)
+## 1.9.1 (2019-10-20)
 
 - Added support for Python 3 from Windows Store (issue [#1162](https://github.com/platformio/platformio-vscode-ide/issues/1162))
 - Better detecting of user home directory on Windows via `%USERPROFILE%` environment variable